remove FormArray<T> from type ControlConfig<T>
FormArray<T> is here plain wrong, because it would create an additional
array dimension. eg. the model has a field with an string array then
a ControlConfig<string[]> -> FormArray<string[]> -> AbstractControl<string[][]>

Normally this API change would be breaking, but it's not possible to build
a working application with this clause. So I will do a patch release.